requirements.txt
livekit>=1.0.13
livekit-agents>=1.2.9
livekit-plugins-openai>=1.2.9
livekit-plugins-silero>=1.2.9
openai>=1.0.0
fastapi==0.104.1
uvicorn==0.24.0
websockets>=15.0.1
pydantic>=2.8.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles>=24.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
